chore: 🐝 Update SDK - Generate 0.0.10

+++ gusto_embedded/src/GustoEmbedded/Models/Requests/GetV1JobsJobIdCompensationsRequest.cs
@@ -26,13 +26,13 @@ namespace GustoEmbedded.Models.Requests
         /// The page that is requested. When unspecified, will load all objects unless endpoint forces pagination.
         /// </summary>
         [SpeakeasyMetadata("queryParam:style=form,explode=true,name=page")]
-        public double? Page { get; set; }
+        public long? Page { get; set; }
 
         /// <summary>
         /// Number of objects per page. For majority of endpoints will default to 25
         /// </summary>
         [SpeakeasyMetadata("queryParam:style=form,explode=true,name=per")]
-        public double? Per { get; set; }
+        public long? Per { get; set; }
 
         /// <summary>
         /// Available options:<br/>
